Fix albumadvisory type bug
If you define a format field ala `%ifdef{albumadvisory}`, assign it
a type of bool, and assign albumadvisory a type of bool, it'll return
True even if the value is False/0 internally. This could be a mismatch
between db type, as I changed the type, but just revert it for now.

Signed-off-by: Christopher Larson <kergoth@gmail.com>