Add 'import_inspect' plugin
This plugin's job is to show candidate/proposed metadata changes before
they're applied. It adds a prompt choice when importing to inspect each
candidate, and has an option (enabled by default) to show the changes to
be applied after you hit Apply, and when timid, will prompt to continue.

Note: if you choose not to continue after hitting apply, it will skip
the current selection and move on to the next, rather than returning to
the candidate menu, so at times it's better to use the menu option to
proactively inspect, and only use the on_apply check as a final sanity
check.

Signed-off-by: Christopher Larson <kergoth@gmail.com>