Only include id in format_{item,album} if it's defined
The only time it's not defined is during the autotagging/import process,
before it's been matched, but adding the ifdef means I can display items
in PromptChoice additions like my `import_inpect` plugin without it
looking ugly. It's not an issue for the `edit` plugin, both because it
doesn't display it with `format()` and because it injects negative ids
to handle mapping the changes.

Signed-off-by: Christopher Larson <kergoth@gmail.com>