Use bold green for text highlight color
I don't want to highlight things with red, as red implies a negative. Also, text_highlight is used by show_model_changes for new fields, and it doesn't make sense to highlight newly added fields as red, but fields changed from the empty string to a value in green, as this is inconsistent.

Signed-off-by: Christopher Larson <kergoth@gmail.com>