Add lsm alias to list albums/items formatted with mediatitles
Franchise and avmedia are also included. For singletons, I don't show
track information at all, as I assume the mediatitle+mediatitledisambig
is sufficient to uniquely identify the track.

Examples, with color disabled:

For a singleton, I see lines like this using `beet lsm`:

    [22798] Quake - Quake (Slipgate OC ReMix) | 2019 | Video Games

For an album, I see lines like this using `beet lsm -a`:

    [2997] Christoph Jordan - Spells & Secrets | 2023 | Video Games
    [3013] Various Artists - Spider-Man - Spider‐Man∶ Into the Spider‐Verse | 2019 | Movies

For an album, I see lines like this using `beet lsm`:

    [22677] Spells & Secrets | 2023 - 01 - Christoph Jordan - The Order Of Things | Video Games

I don't love the albumartist/artist duplication, so may change it to only show if it differs.

Signed-off-by: Christopher Larson <kergoth@gmail.com>